refactor(rl): share sample mask filtering
