+++ nemo_rl/algorithms/ppo.py
@@ -51,6 +51,7 @@ from nemo_rl.algorithms.reward_functions import (
     apply_reward_shaping,
 )
 from nemo_rl.algorithms.utils import (
+    apply_mask_sample_filter,
     print_efficiency_summary,
     print_performance_metrics,
     set_seed,
@@ -280,24 +281,6 @@ def _apply_ppo_seq_logprob_error_masking(
         )
 
     return advantage_mask, metrics
-
-
-def _apply_mask_sample_filter(repeated_batch: BatchedDataDict[DatumSpec]) -> int:
-    """Zero loss_multiplier where mask_sample is True and return the count."""
-    if "mask_sample" not in repeated_batch:
-        return 0
-
-    loss_multiplier = repeated_batch["loss_multiplier"].clone()
-    mask_sample = repeated_batch["mask_sample"]
-
-    if isinstance(mask_sample, list):
-        mask_sample = torch.tensor(mask_sample, dtype=torch.bool)
-    mask_sample_bool = mask_sample.bool()
-
-    num_masked = int(mask_sample_bool.sum().item())
-    loss_multiplier[mask_sample_bool] = 0
-    repeated_batch["loss_multiplier"] = loss_multiplier
-    return num_masked
 
 
 class PPOLoggerConfig(LoggerConfig):
@@ -1593,7 +1576,7 @@ def ppo_train(
                         loss_multiplier[truncated] = 0
                         repeated_batch["loss_multiplier"] = loss_multiplier
 
-                    metrics["num_mask_sample_filtered"] = _apply_mask_sample_filter(
+                    metrics["num_mask_sample_filtered"] = apply_mask_sample_filter(
                         repeated_batch
                     )
 
@@ -2629,7 +2612,7 @@ def async_ppo_train(
                         loss_multiplier[truncated] = 0
                         repeated_batch["loss_multiplier"] = loss_multiplier
 
-                    metrics["num_mask_sample_filtered"] = _apply_mask_sample_filter(
+                    metrics["num_mask_sample_filtered"] = apply_mask_sample_filter(
                         repeated_batch
                     )
